zkp2p-v2-contracts: Initial support for BRL on Mercado Pago
For some reason looks like in Brazilian MP they choose negative values
for outgoing transaction while Argentina uses a positive value for this.

So we need to work around that by pass the currency to the amount
verification function, adjusting for this negative value on BRL payments
and for the comman separator we use (and love).

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>